Bulk Add all helper scripts
In the scripts directory are all the helper scripts for use with makemore. They normally are run from the same directory as makemore.py and the dataset if any. I have put them in scripts folder to remove clutter from the base dir, they can be moved up to base dir, symlink'd or modded to run under the lwoer scripts dir.