Quotes file, sorted
This is the quotes file that is used in the verbose readme. But, it has been sorted unique (sort -u) and outputted to a new file as there are some redundant quotes that I saw in the original file. I noticed this after running training on it and writing about it in the verbose readme.