Expanded on Math example
Appendix 2A example with Math, add,sub,mul div from 0-28, excluding div/0. Did some expansion on the example using a grid search for hyperparameters with constraints on the model size. Included plots of the performance of the model with regards to loss and plots of what the model got right.

Interesting part was that the search revealed a general preference for models with a low (1,2) number of layers and many heads and embeddings. May continue investigating this in the future.